# backend/Models/openAPI.yml
openapi: 3.0.3
info:
  title: Brewtiful API
  version: 1.0.0
servers:
  - url: http://{{baseurl}}
...
paths:
...
  /api/Cart:
    get:
      summary: Get all carts
      deprecated: false
      description: ''
      tags:
        - api/Cart
      parameters:
        - name: Accept
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            '*/*':
              schema:
                type: object
                properties: {}
              examples:
                '1':
                  summary: Get all carts
                  value: "{\r\n    \"_id\": \"integer\",\r\n    \"name\": \"string\",\r\n    \"isSelected\": \"boolean\",\r\n    \"cartItems\": [\r\n        {\r\n            \"id\": \"integer\",\r\n            \"count\": \"integer\",\r\n            \"itemId\": \"integer\",\r\n            \"item\": {\r\n                \"id\": \"integer\",\r\n                \"name\": \"string\",\r\n                \"categoryId\": \"integer\",\r\n                \"category\": {\r\n                    \"id\": \"integer\",\r\n                    \"name\": \"string\"\r\n                },\r\n                \"ingredientNames\": \"array<string>\",\r\n                \"ingredients\": [\r\n                    {\r\n                        \"id\": \"integer\",\r\n                        \"name\": \"string\",\r\n                        \"defaultQuantity\": \"integer\",\r\n                        \"extraCost\": \"float\"\r\n                    },\r\n                    {\r\n                        \"id\": \"integer\",\r\n                        \"name\": \"string\",\r\n                        \"defaultQuantity\": \"integer\",\r\n                        \"extraCost\": \"float\"\r\n                    }\r\n                ],\r\n                \"price\": \"float\"\r\n            },\r\n            \"cartId\": \"integer\",\r\n            \"cart\": \"object\",\r\n            \"modifiedIngredients\": [\r\n                {\r\n                    \"name\": \"string\",\r\n                    \"quantity\": \"integer\"\r\n                }\r\n            ],\r\n            \"totalPrice\": \"float\"\r\n        }\r\n    ],\r\n    \"userId\": \"integer\",\r\n    \"user\": \"object\",\r\n    \"status\": \"string\"\r\n}\r\n{\r\n    \"_id\": \"integer\",\r\n    \"name\": \"string\",\r\n    \"isSelected\": \"boolean\",\r\n    \"cartItems\": [\r\n        {\r\n            \"id\": \"integer\",\r\n            \"count\": \"integer\",\r\n            \"itemId\": \"integer\",\r\n            \"item\": {\r\n                \"id\": \"integer\",\r\n                \"name\": \"string\",\r\n                \"categoryId\": \"integer\",\r\n                \"category\": {\r\n                    \"id\": \"integer\",\r\n                    \"name\": \"string\"\r\n                },\r\n                \"ingredientNames\": \"array<string>\",\r\n                \"ingredients\": [\r\n                    {\r\n                        \"id\": \"integer\",\r\n                        \"name\": \"string\",\r\n                        \"defaultQuantity\": \"integer\",\r\n                        \"extraCost\": \"float\"\r\n                    },\r\n                    {\r\n                        \"id\": \"integer\",\r\n                        \"name\": \"string\",\r\n                        \"defaultQuantity\": \"integer\",\r\n                        \"extraCost\": \"float\"\r\n                    }\r\n                ],\r\n                \"price\": \"float\"\r\n            },\r\n            \"cartId\": \"integer\",\r\n            \"cart\": \"object\",\r\n            \"modifiedIngredients\": [\r\n                {\r\n                    \"name\": \"string\",\r\n                    \"quantity\": \"integer\"\r\n                }\r\n            ],\r\n            \"totalPrice\": \"float\"\r\n        }\r\n    ],\r\n    \"userId\": \"integer\",\r\n    \"user\": \"object\",\r\n    \"status\": \"string\"\r\n}\r\n"
          headers: {}
      security: []
    post:
      summary: Invalid fields (400)
      deprecated: false
      description: ''
      tags:
        - api/Cart
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
        - name: Accept
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                _id:
                  type: integer
                Name:
                  type: string
                IsSelected:
                  type: boolean
                UserId:
                  type: integer
                User:
                  type: object
                  properties:
                    Id:
                      type: integer
                    Name:
                      type: string
                    UserLoginId:
                      type: integer
                    UserLogin:
                      type: object
                      properties:
                        Id:
                          type: integer
                        Username:
                          type: string
                        Password:
                          type: string
                        Token:
                          type: string
                      required:
                        - Id
                        - Username
                        - Password
                        - Token
                    Points:
                      type: integer
                    Role:
                      type: string
                    Carts:
                      type: array
                      items:
                        type: string
                  required:
                    - Id
                    - Name
                    - UserLoginId
                    - UserLogin
                    - Points
                    - Role
                    - Carts
                Status:
                  type: string
                CartItems:
                  type: array
                  items:
                    type: string
              required:
                - _id
                - Name
                - IsSelected
                - UserId
                - User
                - Status
                - CartItems
            example:
              _id: 13
              Name: Domas4's Cart
              IsSelected: true
              UserId: 1
              User:
                Id: 1
                Name: Domas123
                UserLoginId: 1
                UserLogin:
                  Id: 1
                  Username: Domas123
                  Password: newpassword
                  Token: efgh5678
                Points: 200
                Role: Member
                Carts: []
              Status: Active
              CartItems: []
      responses:
        '400':
          description: ''
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: object
                  properties:
                    propertyName:
                      type: string
                    errorMessage:
                      type: string
                    attemptedValue:
                      type: string
                    customState:
                      type: 'boolean'
                    severity:
                      type: integer
                    errorCode:
                      type: string
                    formattedMessagePlaceholderValues:
                      type: object
                      properties:
                        RegularExpression:
                          type: string
                        PropertyName:
                          type: string
                        PropertyValue:
                          type: string
                      required:
                        - RegularExpression
                        - PropertyName
                        - PropertyValue
              examples:
                '1':
                  summary: Invalid fields (400)
                  value: "[\r\n    {\r\n        \"propertyName\": \"Name\",\r\n        \"errorMessage\": \"The name must be '{Name}'s Cart'.\",\r\n        \"attemptedValue\": \"{Name}'s Cart\",\r\n        \"customState\": null,\r\n        \"severity\": 0,\r\n        \"errorCode\": \"RegularExpressionValidator\",\r\n        \"formattedMessagePlaceholderValues\": {\r\n            \"RegularExpression\": \"^{Name}'s Cart$\",\r\n            \"PropertyName\": \"Name\",\r\n            \"PropertyValue\": \"{Name}'s Cart\"\r\n        }\r\n    }\r\n]"
          headers: {}
      security: []
...